+++ src/main/kotlin/dev/mayankmkh/intellij/linear/models/LinearComment.kt
@@ -1,7 +1,7 @@
 package dev.mayankmkh.intellij.linear.models
 
-import apolloGenerated.dev.mayankmkh.intellij.linear.fragment.ShortIssueConnection
 import com.intellij.tasks.Comment
+import dev.mayankmkh.intellij.linear.apolloGenerated.fragment.ShortIssueConnection
 import java.util.Date
 
 class LinearComment(private val node: ShortIssueConnection.Node2) : Comment() {